+++ src/app/app.module.ts
@@ -6,13 +6,14 @@ import { AppComponent } from './app.component';
 import { HomeComponent } from './pages/home/home.component';
 import { DetailComponent } from './pages/detail/detail.component';
 import { NotFoundComponent } from './pages/not-found/not-found.component';
+import { PiechartComponent } from './components/piechart/piechart.component';
 // PrimeNG imports
 import { ChartModule } from 'primeng/chart';
 import { ProgressSpinnerModule } from 'primeng/progressspinner';
 import { MessagesModule } from 'primeng/messages';
 
 @NgModule({
-  declarations: [AppComponent, HomeComponent, DetailComponent, NotFoundComponent],
+  declarations: [AppComponent, HomeComponent, DetailComponent, NotFoundComponent, PiechartComponent],
   imports: [BrowserModule, AppRoutingModule, HttpClientModule, ChartModule, ProgressSpinnerModule, MessagesModule],
   providers: [],
   bootstrap: [AppComponent],
