+++ src/app/components/piechart/piechart.component.html
@@ -1,0 +1,3 @@
+<div class="pie-container">
+    <p-chart #chart type="pie" [data]="data" [options]="options" (onDataSelect)="handleClick($event)" />
+</div>
