+++ src/app/components/piechart/piechart.component.scss
@@ -1,0 +1,9 @@
+:host {
+    .pie-container {
+        text-align: center;
+        margin: 0,5rem;
+        width: 100%;
+        max-width: 650px;
+        margin-top: 2rem;
+    }
+}
